chore(refactor): Reestructure provider component and update imports
