+++ tsconfig.json
@@ -26,9 +26,6 @@
       "@components/*": [
         "./components/*"
       ],
-      "@reducers/*": [
-        "./reducers/*"
-      ],
       "@locales/*": [
         "locales/*"
       ],
